+++ libft/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: marezgui <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/17 15:36:49 by marezgui          #+#    #+#             */
-/*   Updated: 2017/11/25 14:57:41 by marezgui         ###   ########.fr       */
+/*   Updated: 2018/01/13 22:27:28 by marezgui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,20 +19,20 @@ char	*ft_strrchr(const char *s, int c)
 	int		len;
 
 	ptr = NULL;
-	len = ft_strlen(s);
+	len = (int)ft_strlen(s);
 	i = len - 1;
 	while (i >= 0)
 	{
 		if (s[i] == (char)c)
 		{
-			ptr = (char*)s;
+			ptr = (char*)(uintptr_t)s;
 			return (ptr + i);
 		}
 		i--;
 	}
 	if ((char)c == '\0')
 	{
-		ptr = (char*)s;
+		ptr = (char*)(uintptr_t)s;
 		ptr += len;
 	}
 	return (ptr);
