+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: marezgui <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/19 11:13:28 by marezgui          #+#    #+#             */
-/*   Updated: 2017/11/25 14:58:04 by marezgui         ###   ########.fr       */
+/*   Updated: 2018/01/13 03:18:21 by marezgui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,22 +14,22 @@
 
 char	*ft_strtrim(char const *s)
 {
-	int		start;
-	int		len;
-	int		end;
-	char	*t;
+	unsigned	int		start;
+	size_t				len;
+	int					end;
+	char				*t;
 
 	if (!s)
 		return (NULL);
 	start = 0;
-	end = ft_strlen(s) - 1;
+	end = (int)ft_strlen(s) - 1;
 	while (s[start] == ' ' || s[start] == '\n' || s[start] == '\t')
 		start++;
 	if (!s[start])
 		return (ft_strnew(0));
 	while (s[end] == ' ' || s[end] == '\n' || s[end] == '\t')
 		end--;
-	len = (end - start) + 1;
+	len = ((unsigned int)end - start) + 1;
 	t = ft_strsub(s, start, len);
 	if (!t)
 		return (NULL);
